fix(lotto) : 비동기처리로직 콜백함수 호출에서  async await으로 변경
