+++ src/App.js
@@ -1,6 +1,8 @@
 const MissionUtils = require("@woowacourse/mission-utils");
+const util = require("./util");
 const Lotto = require("./Lotto.js");
 const [Console, Random] = [MissionUtils.Console, MissionUtils.Random];
+
 
 class App {
   #totalMoney;
@@ -41,17 +43,24 @@ class App {
   }
 
   readMoney() {
-    Console.readLine("구입금액을 입력해주세요 : ", (money) => {
-      const priceNum = parseInt(money);
-      if (priceNum % 1000 !== 0) {
-        throw new Error("[ERROR] price가 1000으로 나누어 떨어지지 않습니다.");
-      }
-      this.setMoney(priceNum);
-      Console.print(this.#totalMoney);
-      this.buyLotto();
-      this.readWinNums();
-    })
-  }
+    return new Promise((resolve, reject) => {
+      Console.readLine("구입금액을 입력해주세요 : ", (money) => {
+        const priceNum = parseInt(money);
+        if (priceNum % 1000 !== 0) {
+          reject(new Error("[ERROR] price가 1000으로 나누어 떨어지지 않습니다."));
+        }
+        resolve(this.setMoney(priceNum));
+      })})
+      // Console.readLine("구입금액을 입력해주세요 : ", (money) => {
+      //   const priceNum = parseInt(money);
+      //   if (priceNum % 1000 !== 0) {
+      //     throw new Error("[ERROR] price가 1000으로 나누어 떨어지지 않습니다.");
+      //   }
+      //   this.setMoney(priceNum);
+      //   this.buyLotto();
+      //   this.readWinNums();
+      // })
+}
 
   buyLotto() {
     const lottoCount = this.getMoney() / 1000;
@@ -67,22 +76,25 @@ class App {
 
   readWinNums() {
     Console.readLine("당첨번호를 입력해주세요. : ", (winNums) => {
-      const winNumRegex = /\d{1,2},\d{1,2},\d{1,2},\d{1,2},\d{1,2},\d{1,2}/;
-      if (!winNumRegex.test(winNums)) {
-        throw new Error("[ERROR] 당첨번호가 입력형식에 맞지 않습니다.");
-      }
-      const winArray = winNums.split(",").map(Number);
-      Console.print(winArray);
-      if (!winArray.every(num => 1 <= num && num <= 45)) {
-        throw new Error("[ERROR] 당첨번호는 1부터 45까지의 숫자여야 합니다.");
-      }
-      this.setWinNums(winArray);
-      Console.print(this.getWinNums());
+      validateWinNums(winNums);
+      this.readBonusNum();
     })
   }
 
-  play() {
-    this.readMoney();
+  readBonusNum() {
+    Console.readLine("보너")
+
+  }
+
+  async play() {
+    try {
+      await this.readMoney();
+      Console.print(this.getMoney());
+    }
+    catch (err) {
+      Console.print(err.message);
+    }
+
   }
 }
