--- src/tempCodeRunnerFile.js
@@ -1,1 +1,0 @@
-Console.print(this.getLotto());
